refactor: types/apiからtypes/domainに名前を変更
- APIのレスポンス以外にも使用する型だったため，ディレクトリ名を変更
- 関連するimportパスを更新
