+++ src/components/molecules/__test__/PrefectureCheckbox.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
-import { Prefecture } from "../../../types/api/prefecture";
+import { Prefecture } from "../../../types/domain/prefecture";
 import { PrefectureCheckbox } from "../PrefectureCheckbox";
 
 import "@testing-library/jest-dom";
