+++ src/components/organisms/__test__/PrefectureCheckboxList.test.tsx
@@ -2,7 +2,7 @@
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
-import { Prefecture } from "../../../types/api/prefecture";
+import { Prefecture } from "../../../types/domain/prefecture";
 import { PrefectureCheckboxList } from "../PrefectureCheckboxList";
 
 import "@testing-library/jest-dom";
