+++ src/components/templates/DashboardTemplate.tsx
@@ -1,4 +1,4 @@
-import { Prefecture } from "../../types/api/prefecture";
+import { Prefecture } from "../../types/domain/prefecture";
 import { PrefectureCheckboxList } from "../organisms/PrefectureCheckboxList";
 
 export interface DashboardTemplateProps {
